feat: refactor number type (#541)

+++ ckb-indexer/src/main/java/org/nervos/indexer/model/resp/CellCapacityResponse.java
@@ -1,7 +1,6 @@
 package org.nervos.indexer.model.resp;
 
 import com.google.gson.annotations.SerializedName;
-import java.math.BigInteger;
 
 public class CellCapacityResponse {
   @SerializedName("block_hash")
@@ -10,5 +9,5 @@ public class CellCapacityResponse {
   @SerializedName("block_number")
   public int blockNumber;
 
-  public BigInteger capacity;
+  public long capacity;
 }
